# api/openapi.yaml
openapi: 3.1.0
info:
  title: StackPilot API
  version: 0.0.0
  description: StackPilot local control-plane API contract.
servers:
  - url: /
security:
  - bearerAuth: []
  - sessionAuth: []
paths:
...
  /api/v1/workspaces:
    get:
      operationId: listWorkspaces
      summary: List registered workspaces
      responses:
        "200":
          description: Registered workspaces in stable order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
    post:
      operationId: registerWorkspace
      summary: Register the fixed manifest in a workspace root
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWorkspaceRequest"
      responses:
        "201":
          description: Workspace registered with its first valid snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableManifest"
  /api/v1/workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      operationId: getWorkspaceDetail
      summary: Get one consistent workspace management snapshot
      responses:
        "200":
          description: Workspace registration, source, manifest, services, ports, and runtime summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: unregisterWorkspace
      summary: Remove StackPilot catalog data without changing workspace files
      description: Requires the workspace to have no active runtime instance or Operation.
      responses:
        "204":
          description: Workspace unregistered.
        "409":
          $ref: "#/components/responses/Conflict"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/probe:
    post:
      operationId: probeWorkspace
      summary: Probe a workspace root and choose registration or initialization
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceProbeRequest"
      responses:
        "200":
          description: A normal probe state and bounded BAT candidates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceProbe"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableManifest"
  /api/v1/workspace-imports/analyze:
    post:
      operationId: analyzeWorkspaceImport
      summary: Persist a bounded read-only BAT analysis draft
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceAnalyzeRequest"
      responses:
        "201":
          description: Persisted structured import draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceImportDraft"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableManifest"
  /api/v1/workspace-imports/drafts/{draftId}:
    parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          pattern: "^draft_[0-9a-f]{32}$"
    get:
      operationId: getWorkspaceImportDraft
      summary: Get a persisted workspace import draft
      responses:
        "200":
          description: Import draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceImportDraft"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspace-imports/drafts/{draftId}/apply:
    parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          pattern: "^draft_[0-9a-f]{32}$"
      - $ref: "#/components/parameters/CSRF"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      operationId: applyWorkspaceImportDraft
      summary: Queue a durable manifest publication and registration Operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceApplyRequest"
      responses:
        "202":
          description: Import Operation accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRef"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableManifest"
  /api/v1/workspace-imports/drafts/{draftId}/corrections:
    parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          pattern: "^draft_[0-9a-f]{32}$"
      - $ref: "#/components/parameters/CSRF"
    post:
      operationId: correctWorkspaceImportDraft
      summary: Create a validated draft from structured field corrections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceImportCorrectionRequest"
      responses:
        "201":
          description: Corrected persisted draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceImportDraft"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableManifest"
  /api/v1/workspace-imports/operations/{operationId}:
    parameters:
      - $ref: "#/components/parameters/OperationId"
    get:
      operationId: getWorkspaceImportOperation
      summary: Get a pre-registration workspace Operation
      responses:
        "200":
          description: Workspace import Operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceImportOperation"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/operations/{operationId}:
    parameters:
      - name: operationId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/OperationId"
    get:
      operationId: getOperation
      summary: Read an operation and its ordered structured steps
      responses:
        "200":
          description: Persisted operation snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Long-lived local token read only from OS-protected storage by CLI clients.
    sessionAuth:
      type: apiKey
      in: cookie
      name: stackpilot_session
  parameters:
    CSRF:
      name: X-StackPilot-CSRF
      in: header
      required: true
      schema:
        type: string
        minLength: 1
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/WorkspaceId"
...
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/OperationId"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
  schemas:
    ErrorCode:
      type: string
      enum:
        - AUTH_BOOTSTRAP_INVALID
        - EVENT_CURSOR_EXPIRED
        - FEATURE_NOT_ENABLED
        - AUTH_BROWSER_REQUEST_REJECTED
        - AUTH_CAPACITY_REACHED
        - AUTH_SESSION_INVALID
        - AUTH_TOKEN_INVALID
        - CONTROL_PLANE_RESTARTED
        - COMPOSE_CONFIG_INVALID
        - COMPOSE_BUILD_CONFIG_INVALID
        - COMPOSE_BUILD_FAILED
        - COMPOSE_BUILD_TIMEOUT
        - COMPOSE_DISCOVERY_FAILED
        - COMPOSE_INSPECT_FAILED
        - COMPOSE_LOG_FOLLOW_FAILED
        - COMPOSE_LIFECYCLE_INVALID
        - COMPOSE_LIFECYCLE_TIMEOUT
        - COMPOSE_NOT_FOUND
        - COMPOSE_OVERRIDE_CONFLICT
        - COMPOSE_OVERRIDE_INVALID
        - COMPOSE_PREFLIGHT_TIMEOUT
        - COMPOSE_PROJECT_IDENTITY_MISMATCH
        - COMPOSE_PROJECT_NOT_FOUND
        - COMPOSE_START_FAILED
        - COMPOSE_STOP_FAILED
        - CONTAINER_UNHEALTHY
        - COMPOSE_VERSION_UNSUPPORTED
        - DOCKER_DAEMON_UNAVAILABLE
        - DOCKER_NOT_FOUND
        - DOCKER_VERSION_UNSUPPORTED
        - HEALTH_NOT_READY
        - HEALTH_READINESS_TIMEOUT
        - HTTP_BODY_MISMATCH
        - HTTP_STATUS_MISMATCH
        - HTTP_TIMEOUT
        - INTERNAL_ERROR
        - LOG_CURSOR_EXPIRED
        - IDEMPOTENCY_KEY_REUSED
        - MANIFEST_DUPLICATE_KEY
        - MANIFEST_CYCLE_DETECTED
        - MANIFEST_DURATION_INVALID
        - MANIFEST_FILE_TOO_LARGE
        - MANIFEST_HEALTH_TARGET_UNSAFE
        - MANIFEST_CHANGED
        - MANIFEST_MULTIPLE_DOCUMENTS
        - MANIFEST_NOT_REGULAR_FILE
        - MANIFEST_PATH_OUTSIDE_WORKSPACE
        - MANIFEST_PORT_RANGE_INVALID
        - MANIFEST_REFERENCE_NOT_FOUND
        - MANIFEST_SCHEMA_INVALID
        - MANIFEST_SEMANTIC_INVALID
        - MANIFEST_TEMPLATE_INVALID
        - MANIFEST_UNKNOWN_FIELD
        - MANIFEST_YAML_INVALID
        - METHOD_NOT_ALLOWED
        - OPERATION_ALREADY_ACTIVE
        - OPERATION_INVALID_STATE
        - OPERATION_NOT_CANCELLABLE
        - OPERATION_NOT_FOUND
        - PLATFORM_NOT_SUPPORTED
        - PORT_CONFLICT
        - PROCESS_EXITED
        - PROCESS_IDENTITY_MISMATCH
        - PROCESS_START_FAILED
        - PROCESS_STOP_FAILED
        - REQUEST_VALIDATION_FAILED
        - RESOURCE_NOT_FOUND
        - RUNNER_NOT_FOUND
        - RUNNER_PATH_UNSAFE
        - RUNNER_VERSION_CHECK_FAILED
        - RUNNER_VERSION_CHECK_TIMEOUT
        - SECRET_INVALID
        - SECRET_NOT_FOUND
        - SECRET_VERSION_CONFLICT
        - TCP_REFUSED
        - TCP_TIMEOUT
        - SUPERVISOR_EXITED
        - SUPERVISOR_UNAVAILABLE
        - WORKSPACE_ALREADY_REGISTERED
        - WORKSPACE_ID_REQUIRED
        - WORKSPACE_MANIFEST_UNAVAILABLE
        - WORKSPACE_NOT_FOUND
        - WORKSPACE_PATH_INVALID
        - WORKSPACE_SCRIPT_NOT_FOUND
        - WORKSPACE_SCRIPT_OUTSIDE
        - WORKSPACE_SCRIPT_TYPE_UNSUPPORTED
        - WORKSPACE_SCRIPT_ENCODING_UNSUPPORTED
        - WORKSPACE_SCRIPT_TOO_LARGE
        - WORKSPACE_SCRIPT_SYNTAX_UNSUPPORTED
        - WORKSPACE_SCRIPT_DANGEROUS
        - WORKSPACE_SCRIPT_REFERENCE_CYCLE
        - WORKSPACE_IMPORT_INCOMPLETE
        - WORKSPACE_IMPORT_PORT_UNCONFIRMED
        - WORKSPACE_IMPORT_DEPENDENCY_UNCONFIRMED
        - WORKSPACE_IMPORT_SOURCE_CHANGED
        - WORKSPACE_IMPORT_DRAFT_EXPIRED
        - WORKSPACE_MANIFEST_CONFLICT
        - WORKSPACE_MANIFEST_WRITE_FAILED
        - WORKSPACE_EDIT_RUNTIME_ACTIVE
        - WORKSPACE_UNREGISTER_RUNTIME_ACTIVE
        - WORKSPACE_RELINK_SYSTEM_MISMATCH
        - WORKSPACE_SYSTEM_ID_CHANGED
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    OperationId:
      type: string
      pattern: "^op_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
...
    OperationRef:
      type: object
      additionalProperties: false
      required: [operationId, state]
      properties:
        operationId:
          $ref: "#/components/schemas/OperationId"
        state:
          type: string
          enum: [queued, running, cancelling, succeeded, failed, cancelled]
...
    Operation:
      type: object
      additionalProperties: false
      required: [id, workspaceId, systemId, type, state, cancellable, createdAt, steps]
      properties:
        id:
          $ref: "#/components/schemas/OperationId"
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
        systemId:
          $ref: "#/components/schemas/ManifestId"
        type:
          type: string
          enum: [start, stop, restart, service-restart, port-plan, refresh, analyze]
        state:
          type: string
          enum: [queued, running, cancelling, succeeded, failed, cancelled]
        cancellable:
          type: boolean
        cancelRequestedAt:
          type: string
          format: date-time
        errorCode:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          minimum: 0
        steps:
          type: array
          maxItems: 128
          items:
            $ref: "#/components/schemas/OperationStep"
...
    ManifestId:
      type: string
      pattern: "^[a-z][a-z0-9-]{0,62}$"
...
    WorkspaceId:
      type: string
      pattern: "^ws_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
    Digest:
      type: string
      pattern: "^[0-9a-f]{64}$"
    RegisterWorkspaceRequest:
      type: object
      additionalProperties: false
      required:
        - path
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 32767
    Workspace:
      type: object
      additionalProperties: false
      required:
        - id
        - systemId
        - systemName
        - path
        - manifestStatus
        - manifestDigest
        - serviceCount
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/WorkspaceId"
        systemId:
          $ref: "#/components/schemas/ManifestId"
        systemName:
          type: string
        path:
          type: string
        manifestStatus:
          type: string
          enum: [valid, invalid]
        manifestDigest:
          $ref: "#/components/schemas/Digest"
        lastErrorCode:
          $ref: "#/components/schemas/ErrorCode"
        serviceCount:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Workspace"
    WorkspaceProbeRequest:
      $ref: "#/components/schemas/RegisterWorkspaceRequest"
    WorkspaceProbe:
      type: object
      additionalProperties: false
      required: [state, path, candidates]
      properties:
        state:
          type: string
          enum: [ready_to_register, initialization_required]
        path:
          type: string
        candidates:
          type: array
          maxItems: 2000
          items:
            type: object
            additionalProperties: false
            required: [path, size]
            properties:
              path:
                type: string
              size:
                type: integer
                minimum: 0
                maximum: 262144
    WorkspaceAnalyzeRequest:
      type: object
      additionalProperties: false
      required: [path, script]
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 32767
        script:
          type: string
          minLength: 1
          maxLength: 1024
    WorkspaceApplyRequest:
      type: object
      additionalProperties: false
      required: [candidateId]
      properties:
        candidateId:
          type: string
          minLength: 1
          maxLength: 64
    WorkspaceImportCorrectionRequest:
      type: object
      additionalProperties: false
      required: [candidateId, systemName, description, serviceDisplayNames, portPreferred, composeRunning, composeBuild]
      properties:
        candidateId:
          type: string
          minLength: 1
          maxLength: 64
        systemName:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 1024
        serviceDisplayNames:
          type: object
          additionalProperties:
            type: string
            minLength: 1
            maxLength: 128
        portPreferred:
          type: object
          additionalProperties:
            type: integer
            minimum: 1024
            maximum: 65535
        composeRunning:
          type: object
          additionalProperties:
            type: boolean
        composeBuild:
          type: boolean
...
    WorkspaceImportDraft:
      type: object
      additionalProperties: false
      required: [id, state, path, expiresAt, draft]
      properties:
        id:
          type: string
          pattern: "^draft_[0-9a-f]{32}$"
        state:
          type: string
          enum: [active, applied, expired]
        path:
          type: string
        expiresAt:
          type: string
          format: date-time
        draft:
          $ref: "#/components/schemas/WorkspaceImportDraftContent"
    WorkspaceImportDraftContent:
      type: object
      description: Bounded structured draft with relative evidence; never contains BAT source text or service environments.
      additionalProperties: false
      required: [systemId, systemName, analyzedAt, candidates]
      properties:
        systemId:
          $ref: "#/components/schemas/ManifestId"
        systemName:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 1024
        sourceScript:
          type: string
          maxLength: 1024
        sourceDigest:
          $ref: "#/components/schemas/Digest"
        analyzedAt:
          type: string
          format: date-time
        candidates:
          type: array
          minItems: 1
          maxItems: 16
          items:
            $ref: "#/components/schemas/WorkspaceImportCandidate"
...
    WorkspaceImportOperation:
      type: object
      additionalProperties: false
      required: [id, type, state, createdAt, steps]
      properties:
        id:
          $ref: "#/components/schemas/OperationId"
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
        type:
          type: string
          enum: [workspace-import-apply, workspace-edit-apply, workspace-relink-apply]
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        errorCode:
          $ref: "#/components/schemas/ErrorCode"
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          minimum: 0
        steps:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [number, key, state]
            properties:
              number:
                type: integer
                minimum: 1
              key:
                type: string
              state:
                type: string
                enum: [pending, running, succeeded, failed, skipped, cancelled]
              startedAt:
                type: string
                format: date-time
              finishedAt:
                type: string
                format: date-time
              errorCode:
                $ref: "#/components/schemas/ErrorCode"
...
    WorkspaceDetail:
      type: object
      additionalProperties: false
      required: [workspace, source, manifest, services, ports, runtime]
      properties:
        workspace:
          $ref: "#/components/schemas/Workspace"
        source:
          type: object
          additionalProperties: false
          required: [type]
          properties:
            type:
              type: string
              enum: [existing-manifest, bat-import, structured-edit, relinked-manifest]
            entryScript:
              type: string
            sourceDigest:
              $ref: "#/components/schemas/Digest"
            analyzedAt:
              type: string
              format: date-time
        manifest:
          type: object
          additionalProperties: false
          required: [digest, apiVersion, yaml, createdAt]
          properties:
            digest:
              $ref: "#/components/schemas/Digest"
            apiVersion:
              type: string
            description:
              type: string
            yaml:
              type: string
              maxLength: 1048576
            createdAt:
              type: string
              format: date-time
        services:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceServiceDetail"
        ports:
          type: array
          items:
            $ref: "#/components/schemas/WorkspacePortDetail"
        runtime:
          type: object
          additionalProperties: false
          required: [state, activeOperationId]
          properties:
            state:
              type: string
              enum: [stopping, failed, starting, degraded, running, stopped]
            activeOperationId:
              oneOf:
                - $ref: "#/components/schemas/OperationId"
                - type: "null"
    WorkspaceServiceDetail:
      type: object
      additionalProperties: false
      required: [id, displayName, driver, mode, runner, workingDirectory, required, dependsOn, readiness]
      properties:
        id:
          $ref: "#/components/schemas/ManifestId"
        displayName:
          type: string
        driver:
          type: string
          enum: [process, compose]
        mode:
          type: string
          enum: [daemon, oneshot]
        runner:
          type: string
        workingDirectory:
          type: string
        required:
          type: boolean
        dependsOn:
          type: object
          additionalProperties:
            type: string
        readiness:
          type: string
    WorkspacePortDetail:
      type: object
      additionalProperties: false
      required: [name, protocol, conflictPolicy, exposure]
      properties:
        name:
          $ref: "#/components/schemas/ManifestId"
        protocol:
          type: string
          enum: [tcp]
        preferred:
          type: integer
          minimum: 1024
          maximum: 65535
        fallbackRange:
          type: string
        conflictPolicy:
          type: string
        exposure:
          type: string
          enum: [loopback]
...
  responses:
...
    BadRequest:
      description: The request or manifest syntax is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: The request conflicts with current catalog state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableManifest:
      description: The manifest is structurally valid input but cannot be registered safely.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...